Add method for generating a set of backup codes for specific counters
This is useful if a a set of counters is maintained on the server which
correspond to pre-generated backup codes. As codes are used, the server
can remove the stored counter. This method allows the server to simply
pass the remaining set of counters and get the corresponding codes for
verification.