Remove the concept of random number generators
The name was incorrect; the requirement is for random bytes
rather than random numbers. In addition, the separation of
random byte generation and secret generation is not a very
useful abstraction. Better to reduce complexity and move the
responsibility for generating random bytes to the secret
generators themselves as they have very little else to do.