Test highest and lowest allowed composer dependencies
This ensures that as well as testing the current dependencies defined
in the lock file, the build system also tests a version of the library
using the highest available versions and the lowest allowed according
to the version constraints specified in composer.json. This process
gives some confidence that the versions in the package manager are
appropriate and acts as an early warning that an update will break things.