Remove Code value object and Hasher service
The Code value object is unnecessary. A code is really just a string
of a pre-defined length, so having access to the full decimal representation
as well as being able to vary the length of the code after generation are
is not useful.

The Hasher service is likewise unnecessary. The specs define that HMACs
should be used along with defined algorithms. PHP supports this with a
simple function call so abstracting this to a service serves little benefit.