Add some protection around short codes being verified
Verification is done against a number of digits determined by the length
of the supplied code. Unless the client validates the code length
before verifying it, this could lead to insufficiently secure
verification since a 1 digit code would have an approximately 1 in 10
chance of being verified correctly.

This change introduces a minimum code length which is verified by the
library itself.